Added physics convex hull.
Added toggle to dsPhysicsEngine to determine if debugging is enabled. When
enabled, a convex hull may be queried for its faces.

Added optional debug data to dsPhysicsShape, along with a function to
destroy it when the shape is destroyed. May be used to hold model data to
draw the physics scene.