Added dsPhysicsMassProperties.
Contains the inertia, the translate and rotate reference frame for the
inertia, and mass. Functions are provided to create mass properties for
same shapes as are supported by dsPhysicsShape. Mass properties may be
transformed, keeping the transform relative to the shape, shifted, moving
the inertia reference frame but keeping the inertia in the same
location/orientation, and combined.

Convex hull faces are now required in all situations since they will be
needed to compute inertia as a triangle mesh.