Fixed translation for mass properties.
Keep inertia centered around center of mass and store the center of mass
separately. The final shift offset is applied for
dsPhysicsMassProperties_getInertia(), since accumulating shifts doesn't
give the correct results.

Replaced dsPhysicsMassProperties_combine() with
dsPhysicsMassProperties_initializeCombined(). This allows the inertia to be
combined into the proper center of mass since it isn't possible to
transform it back to the center after the fact.

Added dsPhysicsMassProperties_getDecomposedInertia() to decompose the
inertia into a rotation and diagonal, which is used for many physics
libraries. This gets the Jacobi eigenvalues and makes sure the eigenvector
is a right-handed transform.