Use SIMD operations when available for quaternions.
Only do so when fast shuffle is available, i.e. with SSE. Without fast
shuffles, such as on ARM NEON, the savings for SIMD don't exceed the cost
for the moves required to manually shuffle.

Renamed dsQuaternion4_invert to dsQuaternion4_conjugate as it isn't a full
inverse. While this was explained in the comment, this is a clearer name.
The conjugate now uses SIMD operations when possible with a generalized
component negate SIMD function.