Added functions to manage shapes for rigid bodies.
Can add, remove, set the transform on, and finalize shapes for rigid
bodies. Added ID to dsPhysicsShapeInstance to differentiate between shapes
when they share the same underlying shape. (same shape but different
transform)

Mass properties are set when finalizing, either using the default with
optional adjustments (explicit mass and offset point of rotation) or a
provided mass properties. When using the default, computing the mass
properties is skipped if it will never be dynamic. When using the default,
the mass properties are computed by the base Physics library to ensure
consistency between underlying implementations that provide different
degrees of utilities and correctness for various shapes.

Removed mutable mass properties flag and function to set mass properties.
This can be achieved by re-finalizing a rigid body with a mutable shape.