Fixed computation of default rigid body mass properties.
Properly transform the mass properties after initialization from the base
shape. dsPhysicsMassProperties_initializeCombined() will now keep the
input mass properties as-is if there's only one, reducing the amount of
transforms required for the typical case of single-shape rigid bodies.

Removed the mass from dsRigidBodyInit as this is now provided when
finalizing shapes.