Added dsRigidBodyGroup.
Rigid bodies may optionally be created with a group, in which case they are
added/removed from the scene together and some implementations may improve
collision detection performance. This is intended for rigid bodies that are
associated with each-other and in close proximity, such as connected with
constraints.

Added dsDefaultRigidBodyGroup to be used for physics implementations where
the underlying library doesn't support grouping rigid bodies. Currently
PhysX is the only major one that supports native rigid body groups through
PxAggregate.