Added functions to add/remove rigid bodies and groups from a scene.
Rigid bodies are added/removed with an array to encourage grouping changes.
Rigid bodies may optionally be activated on insertion.

Rigid body groups are added/removed individually, and will internally
add/remove the component rigid bodies. The default rigid body group
implementation will call into the physics engine's function pointers with
the internal array of rigid bodies, protected by the spinlock.

Rigid body groups must not be a part of a scene when creating a rigid body
with them. This avoids complications with synchronization. This check isn't
completely foolproof when adding/removing from a scene concurrently, but
the check should be good enough as a sanity check.