Worked around crash in tests on x86 Macs.
When wrapping the curve function in CubicCurveTest, explicitly create a
copy of the point with memcpy for x86 clang builds. When building in
release with clang on x86 (at least for Mac), it somehow gets a null
reference when pushing back on the vector within the lambda. Creating a
copy with memcpy works around this, as the explicit memcpy call avoids the
optimization that triggers the bug.