Added functions to get material for a rigid body.
Added function pointer in shape type to get material. Added fields to
dsPhysicsMesh to get the material for a triangle. Other shape types don't
set the get material function.

Rigid body material gets the shape and calls the get material function if
present. If it's not permitted to get the material (i.e. there is no
material), the rigid body material valaues are used.