Added physics actor contact manifold and callback functions.
Added dsPhysicsActor as a base for dsRigidBody. If soft body support is
added in the future, this will be a base for that as well to be shared for
contact manifold information and callbacks.

Added dsPhysicsActorContactManifold to store the contact points. When
inside the contact properties update callback, the contact properties may
be updated. Depending on the underlying implementation, it may have one set
of properties for the full manifold or one for each contact point.

Callbacks may be set on a physics scene to update the contact settings and
respond to contact manifolds being added, updated, or removed.