Replaced function to get contact properties to one to get defaults.
Rather than explicitly getting the existing contact properties, the
defaults may be explicitly computed. This may be re-used by
implementations, and has fewer calls to forward to the implementation.

Changed update contact properties function to return a bool for whether the
properties were updated. This can be used to avoid duplicate work in the
implementation when setting properties on the underlying physics library.