Added active flag for dsRigidBody.
This will likely be commonly accessed to determine whether the transform
needs to be updated, such as when integrating with a scene. May now choose
whether to activate a rigid body when changing its position.

Adjusted ordering of dsRigidBody members. User data, active state, and
transform are now kept close to each-other to ensure better cache locality
since they will be the most commonly accessed members.