Added shared physics initialization functions and added scene locking.
Added shared initialization functions for physics shapes, rigid body, and
scene. These initialize the shared members, and in the case of shapes also
populates the bounds.

Removed the public shape destroy functions as they are destroyed when the
reference count reaches 0. The destroy function pointer now takes the
physics engine as the first parameter so the engine destroy functions may
be directly assigned to reduce the number of function pointer calls.

Renamed dsPhysicsSceneLimits to dsPhysicsSceneSettings. This now also holds
a bool for whether multi-threaded modifications may be made and the initial
gravity. When multi-threaded modifications may be made, a dsReadWriteLock
is created. Added functions to perform read and write locking, which work
with a dsPhysicsSceneLock to enforce proper locking. Functions to add and
remove rigid bodies now take a dsPhysicsSceneLock to make sure that the
scene has been locked for writing.

Added function to get actors from a scene. This takes an output buffer to
populate with pointers. An index offset may be used to either work with a
specific subset of actors or make multiple calls when the buffer is too
small for all bodies. A dsPhysicsSceneLock is taken to make sure that the
scene has been locked for reading or writing.

Added dsPhysicsScene_update() to update the physics simulation. This takes
a time step and number of steps to perform the update on. This implicitly
locks for writing, but forwards a read lock to provide to the step update
callbacks. This allows the step updates to query the actors, but not
add/remove actors or constraints.