Added common physics constraint interface.
Store two actor pointers for the constraint.

Each constraint has function pointers to query the last applied force and
torque. These are per-constraint rather than shared in the physics engine
to account for Jolt having separate members and queries for each constraint
type compared to Bullet and PhysX that stores them in the base constraint
classes.

Added common function to set whether a constraint is enabled.