Added dsRevolutePhysicsConstraint.
A revolute constraint revolves along an arbitrary axis and may model a
hinge or axle.

Unified how constraint motors are handled. Added an enum to handle whether
the motor is disabled, position, or velocity. The maximum torque or force
is applied when disabled to stop motion, otherwise the target will be
interpreted as a position or velocity depending on the motor type and the
maximum force or torque is applied to reach the target. In the case of
dsSwingTwistPhysicsConstraint a velocity motor isn't supported.