Added dsThreadObjectStorage.
Stores thread-local storage similar to dsThreadStorage, but allows for
cleanup of objects when threads exit or the dsThreadObjectStorage instance
is destroyed. On Windows this uses fiber-local storage, which supports
this use case directly. On other platforms with pthreads it uses a list to
track the values to clean up when dsThreadObjectStorage is destroyed since
pthread keys don't call the destructor on deletion.

This has more overhead than dsThreadStorage, so it is only intended to be
used when cleanup is required. To enforce it isn't used for trivial cases,
the cleanup callback is required to be non-null.