Allow creation of physics constraints without actors.
Physics constraints may now be cloned, which allows for using constraints
as templates. This allows for constraints to be created with one or more
actors missing to be cloned with the specific actors. Connected constraints
may optionally be provided to handle gear or rack and pinion constraints,
where the revolute or slider constraint may be provided to improve the
accuracy.