Adjusted node relation for dsSceneAnimationTransformNode.
No longer a subtype of dsSceneTransformNode. The base transform is set
directly through dsSceneAnimationlist, and doesn't rely on
dsSceneTransformNode for any of the transform management.

Added the ability to ignore the parent transform in a dsSceneTreeNode. This
may be used in situations like a node hooked up to a dsRigidBody, where the
transforms are intended to be global.