Implemented loading of rigid body nodes.
Can either load as RigidBodyNode, intended to be a part of a
RigidBodyGroupNode, or UniqueRigidBodyNode, intended to be a single
instance in the scene graph.

Fixed loading of scene physics constraints to use dsScenePhysicsConstraint
instance. This includes the base dsPhysicsConstraint and optional instance
names for the rigid bodies and connected constraints.