Automatically detect shader configurations for CMake.
Glob .conf files from the directory rather than using hard-coded lists.
This is especially error-prone for Metal, which has a new version every
year.

Disabled deprecation warnings for PVRTC types in Metal. These are
deprecated as of iOS 18 and macOS 15, but may still be useful for older
devices that don't have support for better formats.

Avoid using Metal feature sets, preferring to use GPU families when
targeting fersions that support the newer API.

Updated default iOS deployment target to 12, which is the minimum with the
latest version of Xcode.