Added DEEPSEA_STATIC_RUNTIME CMake flag.
Allows for linking to the static runtime library on Windows. This is
synchronized with the MSL build to ensure it is consistent. This option is
explicitly not allowed with pre-built libraries as they are built with the
dynamic runtime.

Updated the minimum CMake version to 3.15 on Windows to support
CMAKE_MSVC_RUNTIME_LIBRARY.