Avoid potential hash collisions.
Added dsUniqueNameID as a global registration of name IDs. This uses a
global hash table with a counter to generate the names. Initialization is
implicitly done as part of dsApplication unless it has been implicitly
initialized before. Can either call dsUniqueNameID_create() to create the
ID (or get it if already created) or dsUniqueNameID_get() to get the ID
only if it's already been created.

Unique strings for profiling now uses the string instead of only the hash
to ensure hash collisions won't return the wrong string.