Added stream functions for operations when full seeking isn't supported.
May query the remaining bytes in a stream and restart to the beginning as
separate operations. Both support falling back to seeking when not
explicitly implemented. This is useful for situations such as reading
compressed streams where they can often be implemented even when arbitrary
seeking cannot.

Loading shader modules and textures may now be done with streams that don't
offer full support for seeking.