Added initial portions of dsZipArchive.
dsArchive implementation that can read from .zip archives. Directory
records are read and their metadata stored, allowing for operations such as
checking the status of a path or iterating over a directory. The paths may
then be opened independently, even across threads. zip64 extensions are
supported, and files may either use DEFLATE or be uncompressed. Encryption
is not supported.

Streams to read data are not yet implemented.