Prefer using newer versions of Vulkan.
Request the latest known version of Vulkan. Prefer using the core functions
for extensions rolled into core.

Updated tools, which includes the newer version of ModularShaderLanguage to
support customizing the SPIR-V version. The SPIR-V version is now properly
reported based on versions of Vulkan, where the API version and supported
SPIR-V versions aren't one-to-one. Added configurations for SPIR-V 1.1 -
1.6, and added newer versions of SPIR-V to TestLighting and TestAnimation
to verify it works.