Added dsAndroidArchive.
Supports reading from an Android AssetManager through the archive
interface, allowing for multiple Android asset packages to hold data. For
example, separate texture packages based on device support or optional
addons. Limitations due to the underlying AAssetManager implementation,
namely that nonexistent paths will be treated as directories and iterating
directories will only list files, are documented.

Fixed iterating over a directory with dsResourceStream on Android, where
the logic to choose the asset manager directory vs. filesystem was flipped.
Getting the next entry for a filesystem entry no longer passes the
incorrect pointer. Re-enabled the associated unit tests on Android.

Archives may now be closed with a function through dsFileArchive. This
allows for generic archive management code that also closes and cleans up
the archive when it's finished.

dsPath_combine() will now ignore the leading ./ (or . if by itself) for the
second path if the first path is non-empty. This will work better for
archives where you want to have a clean path, and in some situations (such
as getting a resource path) require a non-empty path.

dsResourceStream will now strip out a leading . or ./ from paths when going
through the Android AssetManager. Allows files to be opened with a leading
. or opening the root directory.