Added support for loading from archives.
Added separate loadArchive() functions for loading directly from archives.

Scene resource references may now be specified as a relative path. This
will use the same type of stream as was originally created, as either a
file, resource, or within the same archive. The path info is now passed
through the Python functions and callbacks through the C loading functions
to ensure the relative paths can be managed.