Added scene handoff node and item list.
Allows for a smooth, interpolated handoff when re-parenting to a different
node. For example, a character picking up an item or handing an item to
another character, resolving any small differences between them. A common
ancestor is found during the reparent operation so relative movement is
preserved during the interpolation, especially in siutations such as moving
the scene origin for floating-point precision reasons.