Automatically choose between Wayland and X11 on Linux.
Uses WAYLAND_DISPLAY environment variable to decide which to use.

Pass display handle when creating window surfaces and hint sizes when
creating and updating render surfaces. This allows for creation on Wayland,
which requires the same surface instance that the window was created with
and connecting to the default display multiple times doesn't return the
same instance. The sizes are required as there is no way to query the size
from the window itself.

Force X11 on OpenGL. While EGL technically supports Wayland, there are two
showstopper issues:
1. PBuffers are not supported, at least by Mesa, and looks to be
   fundamentally impossible. This may make it impossible to create a dummy
   surface, as the only other option would be to create a window which will
   likely be visible.
2. The display pointer will need to be provided, but in frameworks like SDL
   are provided with the window. This causes a cyclical requirement where
   the window is needed to create the renderer, but the renderer is required
   to create the appilcation and window.

For now, Vulkan is required for native Wayland.

Updated to the latest OpenGL version and extensions. Also fixed a build
issue with EGL on some platforms with NULL not defined.