Added initial code to support OpenGL on Wayland.
Platform support for OpenGL is now abstracted at runtime rather than
conditionally compiled to the same function names based on the desired
loader, allowing for the loader (namely GLX and EGL) to be chosen at
runtime.. Functions for creating background surfaces have been added to
dsRendererOptions, allowing for Wayland's lack of support for either
PBuffer or Pixmap surfaces.

Added dsSDLApplication_prepareRendererOptions() function to modify the
renderer options as needed based on the renderer ID. In the case if
Wayland, it will initialize the driver, get the display pointer, and set
functions to create hidden windows as background surfaces. This is
currently disabled since an egl_window must be created to create an EGL
surface, and in SDL2 creating the egl_window will also create the EGL
surface with no way to access it. SDL3 allows setting a property to create
the egl_window without creating the EGL surface, so this code path will be
enabled once migrated to SDL3.