Added dsSceneShiftNode.
Allows for having an origin that the sub-graph is relative to. The
immediate children will be shifted with the shiftNodeFunc() function, which
in the case of transform nodes will adjust the transform. dsPhysicsList
can also have a shift node added to it, in which case it will adjust the
rigid body positions when the origin changes.