Added dsSceneAnimationRagdollNode.
Reads the transform from the scene graph based on a relative node a number
of nodes up in the tree to contribute to a direct animation. Two ragdoll
types are currently supported: skeleton, intended to conditionally drive
the main skeleton, and addition, intended for parts on top of the main
skeleton always (or almost always) driven by physics. These can be updated
with dsSceneAnimationList_updateRagdolls(), intended to be done as part of
the physics post-step update.

Added a helper function to get the current transform of a dsSceneTreeNode
relative to an ancestor. This will cut off multiplying the parent
transforms once the ancestor is reached, unless it hits a "no parent"
transform that requires multiplying by the inverse of the ancestor.

Fixed computing the current transform for a dsSceneTreeNode, where
previously a "no parent" transform would always be identity.

The tree node is now provided when removing nodes from scene item lists.
This allows differentiation for the node type when combined within the same
entry, such as the ragdoll nodes tracked within animation node entries.