Added functions to hash and check equality of dsSceneItemList.
Common values are handled by base functions, and function pointers may be
set for any custom values. If no custom values are used, the functions may
be unset. A similar approach is also used for dsSceneInstanceData and
dsSceneInstanceVariables, which is used internally for some dsSceneItemList
types. These may be used to determine if a dsSceneItemList is equivalent to
an existing one within a hash table.