Added support to create a scene based on another scene.
This will move over the nodes for the scene graph and check for equivalent
scene item lists. When an equivalent scene item list is found, it will be
used in place of the new one to avoid having to re-build the entry list.
This will also preserve any internal state, so item lists where this is
important (e.g. animations, physics) can be ensured to remain the same
across all scene configurations to avoid having to manually persist the
state.