Simplified setup for TestLighting.
Combined materials and views into a single set. At runtime, scenes are now
swapped out by loading the new scene and re-using the previous scene for
the item lists where possible.

Added the ability to re-assign a scene on a view. This allows for the use
case of re-creating a scene, such as to change graphics quality settings,
without having to fully re-create the view as well. The allocator is now
the first parameter for view creation and loading rather than second, and
it no longer defaults to the scene's allocator, as the scene is no longer
immutable.