Fixed invalid usage of debug commands in Vulkan.
Avoid debug group and query commands within a render subpass using
secondary command buffers on Vulkan. For built-in GPU profiling, this will
use the framebuffer for the debug group and avoid debug groups for
subsequent subpasses. If the first subpass was not secondary but a later
subpass is, that subpass' debug group will be skipped.

The original behavior is used for other renderer types.