Fixed inconsistent font rendering between qualities.
Fixed texture coordinates so it no longer gets smaller with higher
qualities or shift in position.

Fixed scaling perfomred for emboldened widths for line break computation
and padding added to text geometry. These sizes are consistent even if the
glyph needs to be scaled within the texture, and the computations weren't
consistent with how the textures were managed. The math happened to work
out that the scales were almost always 1 or very close to it, which hid
most potential errors.