Allow depth-only framebuffers on OpenGL.
This was a restriction not necessary so long as glDrawBuffers() may be
called, which is in all cases other than GLES 2.0. TestLighting and
TestAnimation no longer create color renderbuffers for shadow passes as
they are no longer needed on any supported platform.

Removed call to glDrawBuffer() as there is no case where glDrawBuffer() is
supported but glDrawBuffers() is not.