Adjusted parameters for text rendering.
Reduced window size from 4 to 2. This allows for more pixels to be
available for the glyphs while still having plenty of range for embolding
and outlines. Embolding and outline values are now adjusted so that a value
of 1 is usable as "very thick but readable". Outline position is now also
centered around 0, which is more intuitive and allows its scale to be
consistent.

Fixed glyphs to properly use the full texture range available for it rather
than sacrificing one texel for padding.

Fixed gounting of glyphs for vector text when creating the draw buffer,
which could cause a crash from not enough room for the text.