Improved quality management of vector images.
When loading a vector image with a target size, the lowest change in
magnitude between the X and Y scale is used to compute the pixel size. This
avoids situations such as a square target size and rectangular image
causing extreme pixel size values. For example, some of the text samples
in TestVectorDraw had almost no anti-aliasing due to being much wider than
short, with an extreme scale to reach the target size impacting the quality
adjustment too much.