Hooked up icons to text rendering.
Icons are now added to the temporary buffer in dsTextRenderBuffer, which
now takes a maximum number of icon glyphs in its creation parameters. Which
icon pointer is also saved for each icon, so they can be batched both
during commit to prepare for drawing and in draw. Separate draw functions
are used for standard vs. icon glyphs as the shaders bound for standard
glyphs will be separate compared to the internal rendering for icons.

Glyphs that have no geometry now have zero-sized bounding boxes set to skip
same as whitespace. Avoids situations where unexpected floating-point
values may be drawn.